+++ src/main/java/com/motkur/thebigone/Model/Group.java
@@ -1,8 +1,5 @@
 package com.motkur.thebigone.Model;
 
-import com.motkur.thebigone.Repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-
 import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.Set;
@@ -14,6 +11,9 @@ public class Group {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
     private String name;
+    private String password;
+    @Transient
+    private String passwordConfirm;
     @Column(name = "created_on")
     private LocalDateTime createdOn;
     @ManyToOne
@@ -36,6 +36,22 @@ public class Group {
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getPasswordConfirm() {
+        return passwordConfirm;
+    }
+
+    public void setPasswordConfirm(String passwordConfirm) {
+        this.passwordConfirm = passwordConfirm;
     }
 
     public LocalDateTime getCreatedOn() {
