+++ src/main/java/com/motkur/thebigone/Web/TasksController.java
@@ -1,0 +1,4 @@
+package com.motkur.thebigone.Web;
+
+public class TasksController {
+}
